Move and simplify identity value lookups
